Replaced manage.vue org contacts datatable with org contacts component
